+++ TameableCreatures/Properties/AssemblyInfo.cs
@@ -2,11 +2,12 @@
 
 [assembly: AssemblyCompany("TameableCreatures")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyFileVersion("0.11.1.0")]
-[assembly: AssemblyInformationalVersion("0.11.1")]
+[assembly: AssemblyFileVersion("0.11.2.0")]
+[assembly: AssemblyInformationalVersion("0.11.2")]
 [assembly: AssemblyProduct("TameableCreatures")]
 [assembly: AssemblyTitle("TameableCreatures")]
-[assembly: AssemblyVersion("0.11.1.0")]
+[assembly: AssemblyVersion("0.11.2.0")]
+